requirements.txt
munch==2.5.0
numpy==1.19.2
sacred==0.8.2
scikit-learn==0.24.1
tensorflow-gpu==2.3.0
scikit-learn==0.24.1